login cu facebook

+++ app/src/main/java/com/example/amg_tennis_club_android/HomeActivity.java
@@ -2,6 +2,7 @@ package com.example.amg_tennis_club_android;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
@@ -42,11 +43,20 @@ public class HomeActivity extends AppCompatActivity {
         button_logout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                SharedPreferences.Editor editor = sharedPreferences.edit();
-                editor.clear();
-                editor.commit();
-                finish();
-                Toast.makeText(HomeActivity.this, "LOGOUT SUCCESSFULLY", Toast.LENGTH_SHORT).show();
+                if(name == null && password == null) {
+                    startActivity(new Intent(HomeActivity.this, MainActivity.class));
+                    finish();
+                    Toast.makeText(HomeActivity.this, "LOGOUT SUCCESSFULLY", Toast.LENGTH_SHORT).show();
+                   // overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right);
+                }
+                else {
+                    SharedPreferences.Editor editor = sharedPreferences.edit();
+                    editor.clear();
+                    editor.commit();
+                    finish();
+                    Toast.makeText(HomeActivity.this, "LOGOUT SUCCESSFULLY", Toast.LENGTH_SHORT).show();
+                 //   overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right);
+                }
             }
         });
